pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-mcp-server"
version = "1.0.0"
description = "MCP Server for Invoice Management System - AI Agent Architecture"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Student", email = "student@example.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "aiosqlite>=0.19.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[project.scripts]
invoice-mcp = "invoice_mcp_server.cli:main"
invoice-server = "invoice_mcp_server.server:run_server"

[project.urls]
Homepage = "https://github.com/student/invoice-mcp-server"
Documentation = "https://github.com/student/invoice-mcp-server#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/invoice_mcp_server --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
